Improve new data checking and add status only MQTT topic
A timeout after the last MQTT message being received is now used to
trigger te check for changed data.
This greatly reduces the maximum time between a MQTT message triggering
a change and the spaceapi MQTT message being sent.

Added a topic that only sends the status part of the spaceapi body, this
is useful for embedded devices that care about the space open state but
nothing else.